librbd: split ReadRequest
