+++ src/librbd/cache/ReplicatedWriteLog.cc
@@ -17,6 +17,7 @@
 #include "librbd/asio/ContextWQ.h"
 #include "librbd/cache/rwl/ImageCacheState.h"
 #include "librbd/cache/rwl/LogEntry.h"
+#include "librbd/cache/rwl/ReadRequest.h"
 #include "librbd/cache/rwl/Types.h"
 #include <map>
 #include <vector>
